LPD-72056 Support Gallery View for Folder Section
