+++ modules/apps/site/site-cms-site-initializer/src/main/java/com/liferay/site/cms/site/initializer/internal/display/context/ViewFolderSectionDisplayContext.java
@@ -72,6 +72,8 @@ public class ViewFolderSectionDisplayContext extends BaseSectionDisplayContext {
 		return new HashMapBuilder<>().putAll(
 			super.getAdditionalProps()
 		).put(
+			"galleryViewEnabled", true
+		).put(
 			"rootObjectEntryFolderExternalReferenceCode",
 			getRootObjectEntryFolderExternalReferenceCode()
 		).build();
